Change test case generation logic
Currently, the randomised test files produced consist of one or more test cases.
However, in many cases, it may be required to produce exactly one test
case in a test file. So this PR changes the test case generation logic
such that the end user will have to include the no of test cases
required in `input_fn.py`